patch 0.1.0155: Fix GCC warnings.
ticket: N/A

There are multiple changes to fix GCC warnings.

Usage of the several buffer values of the OCRx registers fixed in ATmega8A.

A new NGSpice patch added to load mcusim.cm library by default (if it is
installed from the same directory only).

+++ src/msim_main.c
@@ -362,7 +362,7 @@ int main(int argc, char *argv[])
 		snprintf(LOG, LOGSZ, "Waiting for incoming GDB "
 		         "connections at localhost:%d...", conf.rsp_port);
 		MSIM_LOG_INFO(LOG);
-		MSIM_AVR_RSPInit(&mcu, (int)conf.rsp_port);
+		MSIM_AVR_RSPInit(&mcu, (uint16_t)conf.rsp_port);
 	}
 
 	rc = MSIM_AVR_Simulate(&mcu, 0, mcu.flashend+1, conf.firmware_test);
